+++ CHANGELOG.md
@@ -11,8 +11,10 @@ Changelog follow https://keepachangelog.com/ format.
 *   `ecolab`:
     *   Added: `from etils.lazy_imports import *` alias of `from
         etils.ecolab.lazy_imports import *`.
+    *   Changed: Mutating a lazy_import module mutate the original one. This
+        allow to mutate `builtins` module for example.
 *   `edc`:
-    *    Changed: `.replace` only added once (not in subclasses).
+    *   Changed: `.replace` only added once (not in subclasses).
 
 ## [0.7.1] - 2022-08-09