+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+## [0.8.0] - 2022-09-12
+
 *   `epath`:
     *   Added: `path.stat()`
     *   Fix performance issues for `path.mkdir`.
@@ -134,7 +136,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   Automated github release
 
-[Unreleased]: https://github.com/google/etils/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/google/etils/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/google/etils/compare/v0.7.0...v0.8.0
 [0.7.1]: https://github.com/google/etils/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/google/etils/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/google/etils/compare/v0.5.1...v0.6.0